refactor: api-guide.adoc 수정
