+++ src/docs/asciidoc/api-guide.adoc
@@ -1,4 +1,5 @@
 = WoowaCrew REST API Docs
+:source-highlighter: highlightjs
 :toc: left
 :toclevels: 4
 
@@ -6,7 +7,22 @@
 == 소개
 이 문서는 WoowaCrew API 명세를 작성한 문서입니다.
 
-== Article
+[[Environment]]
+== 개발 환경
+[cols="2*^"]
+|===
+| 환경 | 버전
+
+| Java | 1.8 JDK
+| Spring Boot | 2.1.6.RELEASE
+| DB(MySQL) | 8.0.16
+|===
+
+== 주의
+각 API 요청에는 권한이 필요할 수 있습니다.
+PRECOURSE인 경우 대부분 접근이 불가능합니다.
+
+== Article API
 
 === Article 목록 조회
 - curl 요청
@@ -65,7 +81,7 @@ include::{snippets}/article/delete/path-parameters.adoc[]
 - 응답 예시
 include::{snippets}/article/delete/http-response.adoc[]
 
-== Degree
+== Degree API
 === 기수별 유저수 조회
 - curl 요청
 include::{snippets}/degree/show/curl-request.adoc[]
@@ -89,7 +105,7 @@ include::{snippets}/degree/showDetail/http-response.adoc[]
 - 응답 필드
 include::{snippets}/degree/showDetail/response-fields.adoc[]
 
-== ADMIN
+== ADMIN API
 -- Admin 권한 필요
 
 === 승인된 유저 목록 조회
