+++ src/test/java/woowacrew/common/document/SnippetUtils.java
@@ -1,4 +1,4 @@
-package woowacrew.document;
+package woowacrew.common.document;
 
 import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.restdocs.payload.ResponseFieldsSnippet;
@@ -11,15 +11,15 @@ public class SnippetUtils {
             fieldWithPath("id").type(JsonFieldType.NUMBER).description("id"),
             fieldWithPath("title").type(JsonFieldType.STRING).description("제목"),
             fieldWithPath("content").type(JsonFieldType.STRING).description("게시글"),
-            fieldWithPath("createdDate").type(JsonFieldType.STRING).description("게시글 생성일"),
-            fieldWithPath("lastModifiedDate").type(JsonFieldType.STRING).description("게시글 수정일"),
             fieldWithPath("userResponseDto.id").type(JsonFieldType.NUMBER).description("작성자 id"),
             fieldWithPath("userResponseDto.oauthId").type(JsonFieldType.STRING).description("작성자 github ouath id"),
             fieldWithPath("userResponseDto.nickname").type(JsonFieldType.STRING).description("작성자 닉네임"),
             fieldWithPath("userResponseDto.birthday").type(JsonFieldType.STRING).description("작성자 생일"),
             fieldWithPath("userResponseDto.userRole").type(JsonFieldType.STRING).description("작성자 Role"),
             fieldWithPath("userResponseDto.degreeResponseDto.id").type(JsonFieldType.NUMBER).description("작성자 기수 id"),
-            fieldWithPath("userResponseDto.degreeResponseDto.degreeNumber").type(JsonFieldType.NUMBER).description("작성자 기수")
+            fieldWithPath("userResponseDto.degreeResponseDto.degreeNumber").type(JsonFieldType.NUMBER).description("작성자 기수"),
+            fieldWithPath("createdDate").type(JsonFieldType.STRING).description("게시글 생성일"),
+            fieldWithPath("lastModifiedDate").type(JsonFieldType.STRING).description("게시글 수정일")
     );
 
     public static ResponseFieldsSnippet articleResponseDtosSnippets = responseFields(
@@ -28,15 +28,15 @@ public class SnippetUtils {
             fieldWithPath("articles.[].id").type(JsonFieldType.NUMBER).description("게시글 id"),
             fieldWithPath("articles.[].title").type(JsonFieldType.STRING).description("제목"),
             fieldWithPath("articles.[].content").type(JsonFieldType.STRING).description("게시글"),
-            fieldWithPath("articles.[].createdDate").type(JsonFieldType.STRING).description("게시글 생성일"),
-            fieldWithPath("articles.[].lastModifiedDate").type(JsonFieldType.STRING).description("게시글 수정일"),
             fieldWithPath("articles.[].userResponseDto.id").type(JsonFieldType.NUMBER).description("작성자 id"),
             fieldWithPath("articles.[].userResponseDto.oauthId").type(JsonFieldType.STRING).description("작성자 github ouath id"),
             fieldWithPath("articles.[].userResponseDto.nickname").type(JsonFieldType.STRING).description("작성자 닉네임"),
             fieldWithPath("articles.[].userResponseDto.birthday").type(JsonFieldType.STRING).description("작성자 생일"),
             fieldWithPath("articles.[].userResponseDto.userRole").type(JsonFieldType.STRING).description("작성자 Role"),
             fieldWithPath("articles.[].userResponseDto.degreeResponseDto.id").type(JsonFieldType.NUMBER).description("작성자 기수 id"),
-            fieldWithPath("articles.[].userResponseDto.degreeResponseDto.degreeNumber").type(JsonFieldType.NUMBER).description("작성자 기수")
+            fieldWithPath("articles.[].userResponseDto.degreeResponseDto.degreeNumber").type(JsonFieldType.NUMBER).description("작성자 기수"),
+            fieldWithPath("articles.[].createdDate").type(JsonFieldType.STRING).description("게시글 생성일"),
+            fieldWithPath("articles.[].lastModifiedDate").type(JsonFieldType.STRING).description("게시글 수정일")
     );
 
     public static ResponseFieldsSnippet userResponseDtoListSnippets = responseFields(
